Projects/CSV_Projects/Wine_Suggestion/requirements.txt
langchain
langchain-openai
pinecone
langchain-pinecone
tqdm
pandas
langchain-community
python-dotenv
streamlit